feat: add form submit logic and move TS types to separate file
